+++ tests/data/package/svdummy/__init__.py
@@ -1,0 +1,3 @@
+import dummy
+
+dummy.calc(2, '+', 2)