+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ Unreleased
 ----------
 - Added a Manifest.in
 - Updated project description in the metadata file
+- Broke code up into modules
+- Added some tests
+- Added travis-ci configuration files
+- Added sphinx-apidoc building to package build option
 
 v0.1.0 - 2016-12-15
 -------------------